wnaf

Module traits

Trait definitions.

Macros

impl_window_sizes 🔒

Traits

WindowSize

Allowed wNAF window size: we use this to precompute the window point sizes, because it’s currently not possible to write bounds for them.

WnafGroup

Extension trait on a Group that provides helpers used by [crate::BoxedWnaf].

WnafSize

Size of the wNAF representation: this should be the type-level equivalent of PrimeField::NUM_BITS + 1, which includes an extra entry for any remaining carry.