wnaf

Trait WindowSize

pub trait WindowSize: Unsigned {
    type TableSize: ArraySize;
}

Allowed wNAF window size: we use this to precompute the window point sizes, because it’s currently not possible to write bounds for them.

Required Associated Types

type TableSize: ArraySize

Number of precomputed points in the window table: 1 << (Self::USIZE - 2).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.